Fix spawn functions after modularization
