+++ index.html
@@ -55,6 +55,9 @@
     </div>
   </div>
   <script src="constants.js"></script>
+  <script src="spawner.js"></script>
+  <script src="upgrades.js"></script>
+  <script src="renderer.js"></script>
   <script src="script.js"></script>
 
 </body>
