+++ script.js
@@ -140,11 +140,20 @@ const state = {
   gameOver: false,
 };
 
+if (typeof module !== "undefined") {
+  global.state = state;
+}
+
 function setPaused(p) {
   state.paused = p;
   if (overlay) overlay.style.display = p ? "block" : "none";
   if (menuOverlay && p) menuOverlay.style.display = "block";
   if (menuOverlay && !p) menuOverlay.style.display = "none";
+}
+
+if (typeof module !== "undefined") {
+  global.setPaused = setPaused;
+  global.menuOverlay = menuOverlay;
 }
 
 const player = {
@@ -154,77 +163,18 @@ const player = {
   speed: GAME_CONSTANTS.PLAYER_SPEED,
 };
 
-function spawnEnemy(forcedType) {
-  let type = forcedType || "miniom";
-  if (!forcedType) {
-    const weights = GAME_CONSTANTS.ENEMY_SPAWN_WEIGHTS;
-    const entries = [
-      ["miniom", weights.miniom, 0],
-      ["tanker", weights.tanker, 3600],
-      ["voador", weights.voador, 7200],
-      ["troll", weights.troll, 14400],
-    ].filter(([t, _w, frame]) => state.timeFrames >= frame);
-    const totalWeight = entries.reduce((sum, [, w]) => sum + w, 0);
-    let r = Math.random() * totalWeight;
-    for (const [t, w] of entries) {
-      if (r < w) {
-        type = t;
-        break;
-      }
-      r -= w;
-    }
-  }
-
-  const baseStats = GAME_CONSTANTS.ENEMY_BASE_STATS;
-  const hpMult = Math.pow(GAME_CONSTANTS.ENEMY_HP_LEVEL_COEFF, state.level - 1);
-  const speedMult = Math.pow(
-    GAME_CONSTANTS.ENEMY_SPEED_LEVEL_COEFF,
-    state.level - 1
-  );
-  const stats = baseStats[type];
-  const groundY = canvas.height - stats.size - 30;
-  const enemy = {
-    x: canvas.width + 20 + Math.random() * 40,
-    y:
-      type === "voador"
-        ? canvas.height / 2 + (Math.random() * 120 - 60)
-        : groundY,
-    speed: stats.speed * speedMult,
-    hp: Math.floor(stats.hp * hpMult),
-    maxHp: Math.floor(stats.hp * hpMult),
-    size: stats.size,
-    burn: 0,
-    burnPct: 0,
-    slow: 0,
-    slowFactor: 1,
-    knockback: 0,
-    flash: 0,
-    type,
-    flying: type === "voador",
-  };
-  if (type === "voador") {
-    enemy.baseY = enemy.y;
-    enemy.zigzag = Math.random() < 0.5;
-    enemy.angle = Math.random() * Math.PI * 2;
-    enemy.amplitude = 30;
-  }
-  if (type === "miniom") {
-    enemy.baseY = groundY;
-    enemy.vy = 0;
-    enemy.jumpCooldown =
-      Math.floor(Math.random() * GAME_CONSTANTS.MINIOM_JUMP_COOLDOWN_VAR) +
-      GAME_CONSTANTS.MINIOM_JUMP_COOLDOWN_MIN; // 2-5 segundos
-  }
-  if (type === "tanker") {
-    enemy.dashCooldown =
-      Math.floor(Math.random() * GAME_CONSTANTS.TANKER_DASH_COOLDOWN_VAR) +
-      GAME_CONSTANTS.TANKER_DASH_COOLDOWN_MIN;
-    enemy.dashDuration = GAME_CONSTANTS.TANKER_DASH_DURATION;
-    enemy.dashTime = 0;
-  }
-  // trolls não têm habilidades especiais
-  state.enemies.push(enemy);
+if (typeof module !== "undefined") {
+  var { spawnEnemy, spawnTrunk } = require("./spawner.js");
+  var {
+    levelUp,
+    applyElementUpgrade,
+    showGeneralUpgrades,
+    chooseUpgrade,
+    applyStatUpgrade,
+  } = require("./upgrades.js");
+  var { drawGame } = require("./renderer.js");
 }
+
 
 function shootBasic() {
   const angle = Math.atan2(state.mouseY - player.y, state.mouseX - player.x);
@@ -322,19 +272,6 @@ function castE() {
   };
 }
 
-function spawnTrunk(x) {
-  const barrier = {
-    x,
-    y: canvas.height - GAME_CONSTANTS.TRUNK_BASE_Y,
-    width: 120,
-    height: 120,
-    hp: 20,
-    elements: [],
-    image: troncoImg,
-    hostile: true,
-  };
-  state.barriers.push(barrier);
-}
 
 function clampCrosshair() {
   if (!state.crosshair || state.turrets.length === 0) return;
@@ -369,6 +306,11 @@ const generalUpgradesPool = [
   UPGRADE_BULLET_AOE,
 ];
 
+if (typeof module !== "undefined") {
+  global.elementOptions = elementOptions;
+  global.generalUpgradesPool = generalUpgradesPool;
+}
+
 const comboMap = {
   "Fire,Ice": "Vapor",
   "Fire,Wind": "Fogo Selvagem",
@@ -394,104 +336,6 @@ function getComboName(elems) {
   return comboMap[key] || null;
 }
 
-function levelUp() {
-  state.xp -= state.xpToNext;
-  state.level++;
-  state.xpToNext = Math.floor(state.xpToNext * GAME_CONSTANTS.XP_LEVEL_COEFF);
-
-  const nextSkill = ["Q", "W", "E"].find((s) => !state.skillsUnlocked[s]);
-  if (nextSkill) {
-    state.skillsUnlocked[nextSkill] = true;
-    return;
-  }
-
-  const opts = [];
-  while (opts.length < 3) {
-    const rand =
-      generalUpgradesPool[
-        Math.floor(Math.random() * generalUpgradesPool.length)
-      ];
-    if (!opts.includes(rand)) opts.push(rand);
-  }
-  state.upgradeOptions = opts;
-  state.pendingUpgrade = true;
-  state.upgradeType = "general";
-  setPaused(true);
-  if (menuOverlay) menuOverlay.style.display = "none";
-  showGeneralUpgrades();
-}
-
-function applyElementUpgrade(key) {
-  const up = state.nextUpgrade;
-  if (!up) return;
-  if (up.type === "element") {
-    state.spellElements[key].push(up.element);
-    state.upgrades[key].push(up.element);
-  }
-  state.pendingUpgrade = false;
-  setPaused(false);
-  state.nextUpgrade = null;
-  state.upgradeType = null;
-  document.getElementById("upgradePrompt").style.display = "none";
-}
-
-function showGeneralUpgrades() {
-  const cont = document.getElementById("generalUpgradePrompt");
-  cont.innerHTML = "";
-  state.upgradeOptions.forEach((u, i) => {
-    const div = document.createElement("div");
-    div.className = "upgradeCard";
-    div.textContent = u.desc;
-    div.onclick = () => chooseUpgrade(i);
-    cont.appendChild(div);
-  });
-  cont.style.display = "block";
-}
-
-function chooseUpgrade(idx) {
-  if (state.upgradeType !== "general") return;
-  const up = state.upgradeOptions[idx];
-  applyStatUpgrade(up);
-  state.upgradeOptions = [];
-  state.upgradeType = null;
-  document.getElementById("generalUpgradePrompt").style.display = "none";
-  state.pendingUpgrade = false;
-  setPaused(false);
-  if (state.level % 5 === 0) {
-    const el =
-      elementOptions[Math.floor(Math.random() * elementOptions.length)];
-    state.nextUpgrade = { type: "element", element: el, desc: el };
-    state.pendingUpgrade = true;
-    setPaused(true);
-    if (menuOverlay) menuOverlay.style.display = "none";
-    document.getElementById("upgradeElement").textContent = el;
-    document.getElementById("upgradePrompt").style.display = "block";
-    state.upgradeType = "element";
-  }
-}
-
-function applyStatUpgrade(up) {
-  if (up.prop === "autoFireDelay") {
-    state.autoFireDelay = Math.max(5, state.autoFireDelay + up.value);
-  } else if (up.prop === "baseDamage") {
-    state.baseDamage += up.value;
-  } else if (up.prop === "qDamageBonus") {
-    state.qDamageBonus += up.value;
-  } else if (up.prop === "wBonusHp") {
-    state.wBonusHp += up.value;
-  } else if (up.prop === "eDamageBonus") {
-    state.eDamageBonus += up.value;
-  } else if (up.prop === "qCooldown") {
-    state.qCooldown = Math.max(60, state.qCooldown + up.value);
-  } else if (up.prop === "turretFireDelay") {
-    state.turretFireDelay = Math.max(10, state.turretFireDelay + up.value);
-  } else if (up.prop === "barrierHeight") {
-    state.barrierHeight += up.value;
-  } else if (up.prop === "bulletAOE") {
-    state.bulletAOE += up.value;
-  }
-  state.generalUpgrades.push(up.desc);
-}
 
 function formatTime(frames) {
   const sec = Math.floor(frames / 60);
@@ -654,108 +498,6 @@ function updateHUD() {
   }
 }
 
-function drawGame() {
-  ctx.clearRect(0, 0, canvas.width, canvas.height);
-
-  state.beams = state.beams.filter((b) => {
-    ctx.strokeStyle = b.color;
-    ctx.lineWidth = b.width;
-    ctx.beginPath();
-    ctx.moveTo(player.x, player.y);
-    b.points.forEach((p) => ctx.lineTo(p.x, p.y));
-    ctx.stroke();
-    b.frames--;
-    return b.frames > 0;
-  });
-
-  ctx.fillStyle = "#444";
-  ctx.fillRect(0, canvas.height - 30, canvas.width, 30);
-
-  if (mageImg.complete) {
-    ctx.drawImage(mageImg, player.x - 40, player.y - 40, 80, 80);
-  } else {
-    ctx.fillStyle = "purple";
-    ctx.beginPath();
-    ctx.arc(player.x, player.y, player.radius, 0, Math.PI * 2);
-    ctx.fill();
-  }
-
-  state.enemies.forEach((e) => {
-    let img;
-    if (e.type === "tanker") img = orcFrames[state.orcFrame];
-    else if (e.type === "voador") img = batImg;
-    else if (e.type === "troll") img = trollImg;
-    else img = goblinFrames[state.goblinFrame];
-    if (img.complete) {
-      ctx.drawImage(img, e.x, e.y, e.size, e.size);
-    } else {
-      if (e.type === "tanker") ctx.fillStyle = "brown";
-      else if (e.type === "voador") ctx.fillStyle = "yellow";
-      else if (e.type === "troll") ctx.fillStyle = "darkgreen";
-      else ctx.fillStyle = "green";
-      ctx.fillRect(e.x, e.y, e.size, e.size);
-    }
-    if (e.flash && e.flash > 0 && state.timeFrames % 2 === 0) {
-      ctx.fillStyle = "rgba(255,0,0,0.4)";
-      ctx.fillRect(e.x, e.y, e.size, e.size);
-      e.flash--;
-    }
-    const hpPct = Math.max(0, e.hp) / e.maxHp;
-    ctx.fillStyle = "red";
-    ctx.fillRect(e.x, e.y - 6, e.size, 4);
-    ctx.fillStyle = "lime";
-    ctx.fillRect(e.x, e.y - 6, e.size * hpPct, 4);
-    if (e.burn > 0 && state.timeFrames % 20 < 10) {
-      ctx.fillStyle = "rgba(255,100,0,0.5)";
-      ctx.fillRect(e.x, e.y, e.size, e.size);
-    }
-  });
-
-  state.bullets.forEach((b) => {
-    if (magiaImg.complete) {
-      ctx.drawImage(magiaImg, b.x - 10, b.y - 10, 20, 20);
-    } else {
-      ctx.fillStyle = b.color || "white";
-      ctx.beginPath();
-      ctx.arc(b.x, b.y, 4, 0, Math.PI * 2);
-      ctx.fill();
-    }
-  });
-
-  state.barriers.forEach((b) => {
-    if (b.image && b.image.complete) {
-      ctx.drawImage(b.image, b.x, b.y, b.width, b.height);
-    } else {
-      ctx.fillStyle = "blue";
-      ctx.fillRect(b.x, b.y, b.width, b.height);
-    }
-  });
-
-  ctx.fillStyle = "orange";
-  state.turrets.forEach((t) => {
-    ctx.beginPath();
-    ctx.arc(t.x, t.y, 10, 0, Math.PI * 2);
-    ctx.fill();
-    if (state.crosshair) {
-      ctx.strokeStyle = "rgba(0,255,255,0.3)";
-      ctx.lineWidth = 4;
-      ctx.beginPath();
-      ctx.moveTo(t.x, t.y);
-      ctx.lineTo(state.crosshair.x, state.crosshair.y);
-      ctx.stroke();
-    }
-  });
-
-  if (state.crosshair) {
-    ctx.drawImage(
-      crosshairImg,
-      state.crosshair.x - state.crosshair.radius,
-      state.crosshair.y - state.crosshair.radius,
-      state.crosshair.radius * 2,
-      state.crosshair.radius * 2
-    );
-  }
-}
 
 function applyElementEffects(enemy, elements) {
   if (!elements) return;
@@ -807,7 +549,7 @@ function updateGame() {
     shootBasic();
   if (!state.paused) {
     if (state.spawnTimer-- <= 0) {
-      spawnEnemy();
+      spawnEnemy(state, canvas, GAME_CONSTANTS);
       state.spawnTimer = state.spawnInterval;
     }
     if (--state.spawnIncreaseTimer <= 0) {
@@ -881,7 +623,8 @@ function updateGame() {
         remainingEnemies.push(e);
       }
     } else {
-      if (e.type === "troll") spawnTrunk(e.x);
+      if (e.type === "troll")
+        spawnTrunk(state, canvas, GAME_CONSTANTS, troncoImg, e.x);
       registerKill(e.type);
       state.xp += GAME_CONSTANTS.XP_PER_ENEMY;
     }
@@ -944,7 +687,8 @@ function updateGame() {
         }
         if (e.hp <= 0) {
           state.enemies.splice(i, 1);
-          if (e.type === "troll") spawnTrunk(e.x);
+          if (e.type === "troll")
+            spawnTrunk(state, canvas, GAME_CONSTANTS, troncoImg, e.x);
           registerKill(e.type);
           state.xp += GAME_CONSTANTS.XP_PER_ENEMY;
         }
@@ -1133,7 +877,9 @@ if (typeof module === "undefined") {
 
   const btn = document.getElementById("spawnTrollBtn");
   if (btn) {
-    btn.addEventListener("click", () => spawnEnemy("troll"));
+    btn.addEventListener("click", () =>
+      spawnEnemy(state, canvas, GAME_CONSTANTS, "troll")
+    );
   }
 
   const restart = document.getElementById("restartBtn");
